chore: update navbar design
